scripts/requirements-kafka-plot.txt
# Optional: consume engine trade stream and plot (pip install -r requirements-kafka-plot.txt)
kafka-python>=2.0.2
matplotlib>=3.8.0
